# docs/swagger.yaml
definitions:
  responses.FailureBody:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  responses.SuccessBody:
    properties:
      data: {}
      success:
        type: boolean
    type: object
  schemas.LoginSchema:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    type: object
  schemas.RegisterSchema:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      mobile_number:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email/username and password, and returns
        a JWT token
      parameters:
      - description: User login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.LoginSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.FailureBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.FailureBody'
      summary: User login
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.RegisterSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
swagger: "2.0"